Rename `nsutil` to `namespaces` to be more Go-like.
